Separate extragradient optimizers and utility functions
