+++ cooper/optim/__init__.py
@@ -1,0 +1,2 @@
+from .extra_optimizers import ExtraAdam, ExtraSGD
+from .utils import partial_optimizer, partial_scheduler
